requirements.txt
streamlit
yfinance
pandas
pandas_datareader
matplotlib
requests
python-dateutil
